backend/requirements.txt
fastapi
uvicorn
python-dotenv
pypdf
pytesseract
Pillow
langchain
langchain_groq
